Refactor skill system context exports
