+++ src/hooks/SkillSystemContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from "react";
+
+import { type SkillSystemContextValue } from "./useSkillSystem.types";
+
+export const SkillSystemContext = createContext<SkillSystemContextValue | undefined>(
+  undefined
+);
