+++ src/hooks/useSkillSystem.tsx
@@ -1,180 +1,7 @@
-import {
-  createContext,
-  useCallback,
-  useContext,
-  useEffect,
-  useMemo,
-  useState,
-  type ReactNode
-} from "react";
+import { useContext } from "react";
 
-import { supabase } from "@/integrations/supabase/client";
-import { type PlayerSkills, useGameData } from "./useGameData";
-import {
-  type SkillDefinitionRecord,
-  type SkillProgressRecord,
-  type SkillRelationshipRecord,
-  type SkillSystemContextValue,
-  type UpdateSkillProgressInput
-} from "./useSkillSystem.types";
-
-const SkillSystemContext = createContext<SkillSystemContextValue | undefined>(undefined);
-
-const toNumber = (value: unknown): number => {
-  if (typeof value === "number") return value;
-  if (typeof value === "string") {
-    const parsed = Number(value);
-    return Number.isFinite(parsed) ? parsed : 0;
-  }
-  return 0;
-};
-
-export const SkillSystemProvider = ({ children }: { children: ReactNode }) => {
-  const { selectedCharacterId, updateSkills } = useGameData();
-
-  const [definitions, setDefinitions] = useState<SkillDefinitionRecord[]>([]);
-  const [relationships, setRelationships] = useState<SkillRelationshipRecord[]>([]);
-  const [progress, setProgress] = useState<SkillProgressRecord[]>([]);
-  const [staticLoading, setStaticLoading] = useState(false);
-  const [progressLoading, setProgressLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-
-  useEffect(() => {
-    let isMounted = true;
-    const loadStaticData = async () => {
-      setStaticLoading(true);
-      setError(null);
-      try {
-        // Note: skill_definitions and skill_relationships tables not implemented yet
-        if (!isMounted) return;
-
-        setDefinitions([]);
-        setRelationships([]);
-      } catch (err) {
-        console.error("Error loading skill system data:", err);
-        if (isMounted) {
-          setError(err instanceof Error ? err.message : "Failed to load skill system data.");
-        }
-      } finally {
-        if (isMounted) {
-          setStaticLoading(false);
-        }
-      }
-    };
-
-    void loadStaticData();
-
-    return () => {
-      isMounted = false;
-    };
-  }, []);
-
-  const refreshProgress = useCallback(async () => {
-    if (!selectedCharacterId) {
-      setProgress([]);
-      return;
-    }
-
-    setProgressLoading(true);
-    setError(null);
-
-    try {
-      // Note: skill_progress table not implemented yet
-      setProgress([]);
-    } catch (err) {
-      console.error("Error fetching skill progress:", err);
-      setError(err instanceof Error ? err.message : "Failed to load skill progress.");
-      setProgress([]);
-    } finally {
-      setProgressLoading(false);
-    }
-  }, [selectedCharacterId]);
-
-  useEffect(() => {
-    void refreshProgress();
-  }, [refreshProgress]);
-
-  const progressBySlug = useMemo(() => {
-    return progress.reduce<Record<string, SkillProgressRecord>>((acc, entry) => {
-      if (entry.skill_slug) {
-        acc[entry.skill_slug] = entry;
-      }
-      return acc;
-    }, {});
-  }, [progress]);
-
-  const updateSkillProgress = useCallback(
-    async ({ skillSlug, newSkillValue, xpGain, timestamp, markUnlocked = false }: UpdateSkillProgressInput) => {
-      if (!selectedCharacterId) {
-        throw new Error("No active character selected.");
-      }
-
-      if (!skillSlug) {
-        throw new Error("A skill slug is required to update skill progress.");
-      }
-
-      const isoTimestamp = timestamp ?? new Date().toISOString();
-      const previous = progressBySlug[skillSlug];
-      const previousTotal = toNumber(previous?.total_xp);
-      const nextTotal = xpGain > 0 ? previousTotal + xpGain : previousTotal;
-      const unlockedAt = markUnlocked
-        ? isoTimestamp
-        : previous?.unlocked_at ?? (xpGain > 0 ? isoTimestamp : null);
-
-      const payload = {
-        profile_id: selectedCharacterId,
-        skill_slug: skillSlug,
-        current_value: newSkillValue,
-        total_xp: nextTotal,
-        last_trained_at: isoTimestamp,
-        updated_at: isoTimestamp,
-        unlocked_at: unlockedAt ?? previous?.unlocked_at ?? null
-      } satisfies Partial<SkillProgressRecord> & { profile_id: string; skill_slug: string };
-
-      // Note: skill_progress table not implemented yet
-      console.log("Would update skill progress:", payload);
-
-      const updatedProgress: SkillProgressRecord = {
-        ...(previous ?? { id: `${selectedCharacterId}-${skillSlug}` }),
-        ...payload
-      };
-
-      setProgress(prev => {
-        const index = prev.findIndex(entry => entry.skill_slug === skillSlug);
-        if (index === -1) {
-          return [...prev, updatedProgress];
-        }
-
-        const next = [...prev];
-        next[index] = { ...next[index], ...updatedProgress };
-        return next;
-      });
-
-      await updateSkills({
-        [skillSlug]: newSkillValue,
-        updated_at: isoTimestamp
-      } as Partial<PlayerSkills>);
-
-      return updatedProgress;
-    },
-    [progressBySlug, selectedCharacterId, updateSkills]
-  );
-
-  const value = useMemo<SkillSystemContextValue>(
-    () => ({
-      definitions,
-      relationships,
-      progress,
-      loading: staticLoading || progressLoading,
-      error,
-      refreshProgress,
-      updateSkillProgress
-    }),
-    [definitions, relationships, progress, staticLoading, progressLoading, error, refreshProgress, updateSkillProgress]
-  );
-
-  return <SkillSystemContext.Provider value={value}>{children}</SkillSystemContext.Provider>;
-};
+import { SkillSystemContext } from "./SkillSystemContext";
+import { type SkillSystemContextValue } from "./useSkillSystem.types";
 
 export const useSkillSystem = (): SkillSystemContextValue => {
   const context = useContext(SkillSystemContext);
